fix: 좌표계 문제 해결 및 로그 제거
