+++ src/main/resources/config/gis.yml
@@ -3,6 +3,3 @@ gis:
     force-xy: true
     source-epsg: ${GIS_SOURCE_EPSG}
     target-epsg: ${GIS_TARGET_EPSG}
-  system-properties:
-    org.geotools.referencing.forceXY: "true"
-    org.geotools.referencing.caching.enabled: "true"
